[omnibox][post-ac] Split features.
Currently, we have 1 umbrella feature with multiple params that would
control 4 batches of experiments.

This was problematic since we decided to run the batches in parallel in
separate studies. This CL splits the feature into 3. (Batch 4 isn't
implemented yet, it will get it's own feature when it is.)

The features & params are:
 - kAutocompleteStability (original umbrella feature)
   - kAutocompleteStabilityDontCopyDoneProviders
   - kAutocompleteStabilityAsyncProvidersFirst
 - kPreserveDefault (batch 2)
   - kAutocompleteStabilityPreserveDefaultAfterTransfer
   - kAutocompleteStabilityPreserveDefaultForSyncUpdatesMinInputLength
   - kAutocompleteStabilityPreserveDefaultForAsyncUpdates
 - kUpdateResultDebounce (batch 3)
   - kAutocompleteStabilityUpdateResultDebounceFromLastRun
   - kAutocompleteStabilityUpdateResultDebounceDelay

Bug: 1340291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3848937
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1038463}

+++ components/omnibox/browser/autocomplete_controller.cc
@@ -151,7 +151,7 @@ void RecordMatchDeletion(const AutocompleteMatch& match) {
 // sync/async update. Sync updates may have a minimum input length as well.
 bool ShouldPreserveDefault(bool in_start, size_t input_length) {
   if (in_start) {
-    static const size_t min_input_length =
+    static const int min_input_length =
         OmniboxFieldTrial::
             kAutocompleteStabilityPreserveDefaultForSyncUpdatesMinInputLength
                 .Get();
